+++ Cargo.toml
@@ -10,5 +10,12 @@ authors = ["Jeff Austin <jja6@pdx.edu>, Kamakshi Nagar <kamakshi@pdx.edu>"]
 edition = "2018"
 
 [dependencies]
-rocket = "0.4.2"
 rand = "0.7.0"
+rocket = "0.4.2"
+serde = "1.0"
+serde_derive = "1.0"
+
+[dependencies.rocket_contrib]
+version = "0.4.1"
+default-features = false
+features = ["handlebars_templates"]
