+++ static/index.css
@@ -1,0 +1,32 @@
+html {
+  background-image: url("images/background.png");
+  text-align: center;
+  color: white;
+}
+
+h1 {
+  font-size: 50px;
+}
+
+#formBox {
+  height: 20px;
+  width: 30%;
+}
+
+a {
+  text-decoration: none;
+  color: black
+}
+
+textarea {
+  margin-top: 5%;
+  margin-bottom: 2%;
+  width: 50%;
+  height: 80%;
+  resize: none;
+}
+
+#goBack {
+  width: 10%;
+  height: 30px;
+}
