+++ static/index.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <title>Index</title>
+    <link rel="stylesheet" href="/static/index.css">
+  </head>
+
+  <body>
+    <h1>List of Question </h2>
+      <textarea readonly rows= "20" cols="200">
+        OPTION PROBLEM TYPE
+
+        0)     Translate Random Virtual Address to a Physical Address
+        1)     Translate Random Virtual Address Returned by Malloc() to a Physical Address
+        2)     Calculate Specified Portion through the Stack as a Virtual Address
+        8)     Generate fresh segmented memory model
+        9)     Generate Random Problem
+        10)    Exit
+  </textarea>
+    <form action="/search" method="post" accept-charset="utf-8" action="result.html">
+      <input id="formBox" type="text" name="term" placeholder="Type Here"/>
+      <input type="submit" value="Next">
+    </form>
+
+  </body>
+</html>
