+++ static/result.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <title>Result</title>
+    <link rel="stylesheet" href="/static/index.css">
+  </head>
+
+  <body>
+    <h1>Result</h2>
+    <form action="/search" method="post" accept-charset="utf-8">
+      <input id="formBox" type="text" name="term" placeholder="Type Here">
+      <button id="submitBtn" type="button"><a href="/static/result.html">Submit</a></button>
+    </form>
+    <textarea readonly rows= "20">
+    </textarea>
+
+    <div>
+    <button id="goBack" type="button"><a href="/static/index.html">Go Back</a></button>
+
+    </div>
+  </body>
+
+</html>
