+++ templates/index.hbs
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <title>Index</title>
+    <link rel="stylesheet" href="/static/index.css">
+  </head>
+
+  <body>
+    <h1>Results</h1>
+    <form action="/search" method="post" accept-charset="utf-8">
+      <input id="formBox" type="text" name="term" placeholder="Type Here">
+      <input type="submit" value="Next">
+    </form>
+    {{~> page}}
+
+    <div>
+    <button id="goBack" type="button"><a href="/static/index.html">Go Back</a></button>
+    </div>
+  </body>
+</html>
